refactor : AuthMessage 위치 컨벤션에 맞게 이동
