+++ orury-client/src/main/java/org/fastcampus/oruryclient/auth/converter/message/AuthMessage.java
@@ -1,4 +1,4 @@
-package org.fastcampus.oruryclient.auth.util;
+package org.fastcampus.oruryclient.auth.converter.message;
 
 import lombok.Getter;
